+++ affected-files.config.js
@@ -1,0 +1,5 @@
+module.exports = {
+  missing: [
+    'missing-library',
+  ],
+}